% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/begin.R
\name{begin}
\alias{begin}
\title{Initialise an autobench run.}
\usage{
begin(file, name, quiet = FALSE, tool = "rbenchmark", min.reps = 1,
  max.reps = 100, min.time = 0.5, check = FALSE, unit = "t",
  stop.on.fail = FALSE)
}
\arguments{
\item{file}{Results file.}

\item{name}{Name of autobench run.}

\item{quiet}{Suppress run status.}

\item{tool}{Benchmarking package used by autobench. One of "bench",
"microbenchmark", and "rbenchmark".}

\item{min.reps}{Only applies if tool = "bench". Minimum number of times to
run each expression.}

\item{max.reps}{If tool = "bench", the maximum number of times to run each
expression. If tool = "microbenchmark" or tool = "rbenchmark", the
exact number of times to run each expression.}

\item{min.time}{Only applies if tool = "bench". Minimum time in seconds to
each expression (unless expression has already been run max.reps times).}

\item{check}{Make sure all expressions in a run return equal values.}

\item{unit}{Only applies if tool = "microbenchmark" or tool = "rbenchmark".
Units to show in benchmarking results tables.}

\item{stop.on.fail}{Whether to give an error if one of the expressions gives
an error.}
}
\description{
Initialise an autobench run.
}
\author{
Benjamin Jean-Marie Tremblay
}
