% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/end.R
\name{end}
\alias{end}
\title{Terminate an autobenchR run.}
\usage{
end()
}
\value{
\code{TRUE}, invisibly.
}
\description{
Calling \code{\link[autobenchR:end]{autobenchR::end()}} will terminate an \pkg{autobenchR} session by
printing total runtime and session info (if \code{session.info = TRUE}). Note that
\code{\link[autobenchR:begin]{autobenchR::begin()}} must be called again in order to execute further
benchmarks.
}
\author{
Benjamin Jean-Marie Tremblay
}
