% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/note.R
\name{note}
\alias{note}
\title{Insert notes in an autobench run.}
\usage{
note(..., name = "Note")
}
\arguments{
\item{...}{User notes. The individual elements
will have two newline characters separating them.}

\item{name}{Name of note section.}
}
\value{
\code{NULL}, insivibly.
}
\description{
Insert notes in an autobench run.
}
\author{
Benjamin Jean-Marie Tremblay
}
