% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/begin.R
\name{begin}
\alias{begin}
\title{Initialise an autobenchR run.}
\usage{
begin(file = stdout(), name = NULL, quiet = ifelse(isatty(file),
  TRUE, FALSE), tool = "rbenchmark", min.reps = 1, max.reps = 100,
  min.time = 0.5, check = FALSE, unit = "t", stop.on.fail = FALSE,
  format = ifelse(isatty(file), "txt", "md"),
  session.info = ifelse(isatty(file), FALSE, TRUE))
}
\arguments{
\item{file}{Print results to this file. The default (\code{file = stdout()})
prints results to the terminal.}

\item{name}{Name of \pkg{autobenchR} session.}

\item{quiet}{Suppress run status. By default, is set to \code{TRUE} if \code{file} is a
terminal connection and \code{FALSE} otherwise.}

\item{tool}{Benchmarking package used by autobenchR. One of "bench",
"microbenchmark", and "rbenchmark".}

\item{min.reps}{Only applies if \code{tool = "bench"}. Minimum number of times to
run each expression.}

\item{max.reps}{If \code{tool = "bench"}, the maximum number of times to run each
expression. If \code{tool = "microbenchmark"} or \code{tool = "rbenchmark"}, the
exact number of times to run each expression.}

\item{min.time}{Only applies if \code{tool = "bench"}. Minimum time in seconds to
each expression (unless expression has already been run \code{max.reps} times).}

\item{check}{Make sure all expressions in a run return equal values. Only
valid if \code{tool = "bench"} or \code{tool = "microbenchmark"}.}

\item{unit}{Only applies if \code{tool = "microbenchmark"} or \code{tool = "rbenchmark"}.
Units to show in benchmarking results tables.
See \code{\link[microbenchmark:microbenchmark]{microbenchmark::microbenchmark()}} for acceptable units when
\code{tool = "microbenchmark"}. When \code{tool = "rbenchmark"}, acceptable units are
"ns" (nanoseconds), "us" (microseconds), "ms" (milliseconds),
"s" (seconds), and "min" (minutes). The default, \code{unit = "t"}, when
run with \code{tool = "microbenchmark"} will automatically pick a unit; and
when run with tool = "rbenchmark", seconds are used.}

\item{stop.on.fail}{Whether to give an error if one of the expressions gives
an error, or continue.}

\item{format}{Format of output results. One of "txt" (text) and "md" (markdown).
Defaults to "txt" if \code{file} is a terminal connection.}

\item{session.info}{Print session info after \code{\link[autobenchR:end]{autobenchR::end()}} is
called. By default will only be set to \code{TRUE} if \code{file} is not a
terminal connection.}
}
\value{
\code{NULL}, invisibly.
}
\description{
Generates the settings used by instances of \code{\link[autobenchR:run]{autobenchR::run()}}. This
function must be called before any other, and again after \code{\link[autobenchR:end]{autobenchR::end()}}
is called in order to run further benchmarks.
}
\author{
Benjamin Jean-Marie Tremblay
}
