% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/end.R
\name{end}
\alias{end}
\title{Terminate an autobench run.}
\usage{
end()
}
\value{
\code{TRUE}, invisibly.
}
\description{
Calling \code{\link[autobench:end]{autobench::end()}} will terminate an \pkg{autobench} session by
printing total runtime and session info (if \code{session.info = TRUE}). Note that
\code{\link[autobench:begin]{autobench::begin()}} must be called again in order to execute further
benchmarks.
}
\author{
Benjamin Jean-Marie Tremblay
}
