% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{run}
\alias{run}
\title{Run a set of benchmarks.}
\usage{
run(name = NULL, ...)
}
\arguments{
\item{name}{Name representing this set of benchmarks.}

\item{...}{Named expressions to benchmark.}
}
\value{
\code{FALSE} if an error occurred (and \code{stop.on.fail = FALSE}), \code{TRUE} if
run instance was successful, and \code{NULL} if skipped.
}
\description{
Must be run after \code{\link[autobenchR:begin]{autobenchR::begin()}}, and before \code{\link[autobenchR:end]{autobenchR::end()}}.
}
\section{Note}{

All warnings and messages generated by benchmarked expressions are
suppresssed.
}

\author{
Benjamin Jean-Marie Tremblay
}
